Fix uploading to wrong directory and directories not creating properly
Fixes files not uploading to the right directory when running. This would result in the root directory of the Pterodactyl panel having all the files created when the files should be located in respected directories, such as a plugin being in the "plugins/" directory.

Fixes directories not being created properly and results in a broken folder structure.